Combine index and task-lists-element files
