+++ rollup.config.js
@@ -1,7 +1,7 @@
 import pkg from './package.json'
 
 export default {
-  input: 'dist/index.js',
+  input: 'dist/task-lists-element.js',
   output: [
     {
       file: pkg['module'],
