+++ test/karma.config.cjs
@@ -1,7 +1,7 @@
-module.exports = function(config) {
+module.exports = function (config) {
   config.set({
     frameworks: ['mocha', 'chai'],
-    files: [{pattern: '../dist/index.js', type: 'module'}, 'test.js'],
+    files: [{pattern: '../dist/task-lists-element.js', type: 'module'}, 'test.js'],
     reporters: ['mocha'],
     port: 9876,
     colors: true,
